requirements.txt
streamlit==1.36.0
torch==2.3.1
torchvision==0.18.1
Pillow==10.2.0